refactor(AuthorAndDate): extract Author component for SSR
